Refactored component hierarchy
