+++ web/js/components/Saldo/index.jsx
@@ -1,0 +1,4 @@
+import { translate } from 'react-i18next';
+import Saldo from './Saldo';
+
+export default translate()(Saldo);
