+++ web/js/components/stories.jsx
@@ -1,0 +1,1 @@
+require('./Saldo/stories.jsx');
